Aggressively reduce memory to prevent OOM kill in Codespaces
The previous incremental-free approach wasn't enough — peak memory
from individual operations still exceeded the ~4GB Codespace limit.

Three high-impact changes:

1. float32 everywhere: All data arrays (lon, lat, field, diff) are now
   stored as float32 instead of float64, halving memory for every array.
   This includes fielddiff output and plotting helper arrays.

2. Coarsen grids before plotting: The diff grid often has 1M+ cells but
   the output image is only ~1300x600 pixels. We now stride-sample down
   to ~500K cells max for pcolormesh, while keeping full resolution for
   the airport ΔT table lookups.

3. Smaller figure in Codespaces: 10x5 instead of 13x6, plus 100 DPI
   instead of 150 — reduces matplotlib's rendering buffer by ~55%.

Also frees Herbie objects (nwp, rtma) immediately after extracting data.

+++ comparator/fielddiff.py
@@ -8,4 +8,4 @@ def compute_fielddiff(hrrr_t2m: xr.DataArray, rtma_on_hrrr_t2m) -> xr.DataArray:
     """
     h, r = xr.align(hrrr_t2m, rtma_on_hrrr_t2m, join="exact")
     valid = (np.isfinite(h) & np.isfinite(r) & (h > 150) & (h < 330) & (r > 150) & (r < 330))
-    return (h - r).where(valid) * 9/5
+    return ((h - r).where(valid) * np.float32(9/5)).astype("float32")
